+++ src/main/java/com/backend/backend_web/service/CalificacionService.java
@@ -12,6 +12,7 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.backend.backend_web.dto.CalificacionDTO;
+import com.backend.backend_web.entity.Arrendador;
 import com.backend.backend_web.entity.Calificacion;
 import com.backend.backend_web.exception.RegistroNoEncontradoException;
 import com.backend.backend_web.repository.CalificacionRepository;
@@ -97,7 +98,12 @@ public class CalificacionService {
         return CalificacionDTO;
     }
 
-    public void delete(Long id) {
+    public void delete(Long id) throws RegistroNoEncontradoException {
+        Optional<Calificacion> calificacion = repository.findById(id);
+        if (!calificacion.isPresent()) {
+            throw new RegistroNoEncontradoException(
+                    "Calificacion no encontrada con ID " + id + " por lo tanto no se puede eliminar");
+        }
         repository.deleteById(id);
     }
